+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/console.php
@@ -52,6 +52,7 @@ use Symfony\Component\Messenger\Command\FailedMessagesRetryCommand;
 use Symfony\Component\Messenger\Command\FailedMessagesShowCommand;
 use Symfony\Component\Messenger\Command\SetupTransportsCommand;
 use Symfony\Component\Messenger\Command\StatsCommand;
+use Symfony\Component\Messenger\Command\ShowMessagesCommand;
 use Symfony\Component\Messenger\Command\StopWorkersCommand;
 use Symfony\Component\Scheduler\Command\DebugCommand as SchedulerDebugCommand;
 use Symfony\Component\Serializer\Command\DebugCommand as SerializerDebugCommand;
@@ -221,6 +222,14 @@ return static function (ContainerConfigurator $container) {
                 service('messenger.transport.native_php_serializer')->nullOnInvalid(),
             ])
             ->tag('console.command')
+
+        ->set('console.command.messenger_show', ShowMessagesCommand::class)
+            ->args([
+                service('messenger.receiver_locator'),
+            ])
+            ->tag('console.command')
+        ;
+
 
         ->set('console.command.messenger_stats', StatsCommand::class)
             ->args([